requirements.txt
streamlit
seaborn
python-dotenv
pandas
matplotlib
requests